requirements.txt
fastapi==0.92.0
pymongo==4.6.1
uvicorn==0.20.0
python-decouple==3.8
requests==2.28.1
cryptography==41.0.4